is_bigger/mlfunctions.py
import torch
import torch.nn as nn

def ml_is_bigger(int1: int, int2: int) -> bool:
    """
    Returns True if int1 is greater than int2, False otherwise
    :param int1: int
    :param int2: int
    :return: True if int1 > int2, False otherwise
    """
    if int1 == int2:
        return False

    class Net(nn.Module):
        def __init__(self):
            super(Net, self).__init__()
            self.fc1 = nn.Linear(2, 100)
            self.fc2 = nn.Linear(100, 100)
            self.fc3 = nn.Linear(100, 50)
            self.fc4 = nn.Linear(50, 1)

        def forward(self, x):
            x = torch.sigmoid(self.fc1(x))
            x = torch.sigmoid(self.fc2(x))
            x = torch.sigmoid(self.fc3(x))
            x = self.fc4(x)
            return x

    net = Net()
    criterion = nn.MSELoss()
    optimizer = torch.optim.Adam(net.parameters(), lr=0.001)

    # train the neural network for more iterations
    for i in range(10000):
        # generate random pairs of numbers for training
        num1 = torch.randn(1, 1)
        num2 = torch.randn(1, 1)
        input = torch.cat((num1, num2), 1)
        # adjust the target to account for equal numbers
        target = torch.tensor([[1.0 if num1 > num2 else 0.0 if num1 < num2 else 0.5]], dtype=torch.float32)
        optimizer.zero_grad()
        output = net(input)
        loss = criterion(output, target)
        loss.backward()
        optimizer.step()

    # test the neural network
    input = torch.tensor([[int1, int2]], dtype=torch.float32)
    output = net(input)
    return output.item() > 0.5

def ml_is_smaller(int1: int, int2: int) -> bool:
    """
    Returns True if int1 is less than int2, False otherwise
    :param int1: int
    :param int2: int
    :return: True if int1 < int2, False otherwise
    """
    return ml_is_bigger(int2, int1)

def ml_is_equal(int1: int, int2: int) -> bool:
    """
    Returns True if int1 is equal to int2, False otherwise
    :param int1: int
    :param int2: int
    :return: True if int1 == int2, False otherwise
    """
    return not ml_is_bigger(int1, int2) and not ml_is_smaller(int1, int2)

def ml_is_not_bigger(int1: int, int2: int) -> bool:
    """
    Returns True if int1 is not greater than int2, False otherwise
    :param int1: int
    :param int2: int
    :return: True if int1 <= int2, False otherwise
    """
    return not ml_is_bigger(int1, int2)

def ml_is_not_smaller(int1: int, int2: int) -> bool:
    """
    Returns True if int1 is not less than int2, False otherwise
    :param int1: int
    :param int2: int
    :return: True if int1 >= int2, False otherwise
    """
    return not ml_is_smaller(int1, int2)

def ml_is_not_equal(int1: int, int2: int) -> bool:
    """
    Returns True if int1 is not equal to int2, False otherwise
    :param int1: int
    :param int2: int
    :return: True if int1 != int2, False otherwise
    """
    return not ml_is_equal(int1, int2)